+++ src/aioway/_torch.py
@@ -5,13 +5,10 @@ import re
 import typing
 from collections import abc as cabc
 
-import tensordict as td
 import torch
 from torch import _ops
 
 __all__ = [
-    "tdict_rename",
-    "tdict_all_equal",
     "HasParam",
     "TensorFilter",
     "is_leaf_has_grad",
@@ -29,18 +26,6 @@ _ATEN_OPS = re.compile("aten::.+")
 _PRIM_OPS = re.compile("prim::.+")
 _TORCHVISION_OPS = re.compile("image::.+")
 _TORCHCODEC_OPS = re.compile("torchcodec_ns::.+")
-
-
-def tdict_rename(tdict: td.TensorDict, **renames: str):
-    return td.TensorDict({renames.get(key, key): value for key, value in tdict.items()})
-
-
-def tdict_all_equal(left: td.TensorDict, right: td.TensorDict, /):
-    if left.keys() != right.keys():
-        return False
-
-    eq: td.TensorDict = left == right
-    return eq.all()
 
 
 class TensorFilter(typing.Protocol):
